nim_workflows/vision_text_extraction/requirements.txt
matplotlib
requests
pillow
pandas
gradio
openai
tqdm
notebook
